Fixed an issue with reading from the temp file.
This resolves https://github.com/Dickurt/vscode-php-formatter/issues/3, where fixing a file would erase the contents. It was the result of the fs.readFileSync method which handled the file descriptor cursor position poorly. Using a read stream, we can set the file cursor position to the first line when reading from the temp file.

Be aware, however, that this code assumes utf-8 character encoding throughout, which could cause a lot of issues. Hence we are waiting for a feature to be implemented with which extensions can get the a file's character encoding. Feature request here: https://github.com/Microsoft/vscode/issues/11690.

Also added some more debug logging, as well as a convenience method that wraps a string in double quotes.