Added phpformatter.arguments setting.
This new setting deprecates previous settings like phpformatter.level and phpformatter.fixers, because it allows any argument to be set, removing the need to add user settings for each and every setting PHP-CS-Fixer has.

Configuration-related variables in the code are now prepended with underscores to distinguish them from other variables.